# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Riwi Cine API",
    "version": "1.0.0",
    "description": "Documentación generada automáticamente con Swagger para la API de Riwi Cine."
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Servidor local versionado (RN-113)"
    },
    {
      "url": "http://localhost:3000/api",
      "description": "Legacy sin versionar (compatibilidad)"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/movies": {
      "get": {
        "summary": "Obtener todas las películas activas",
        "tags": [
          "Movies"
        ],
        "responses": {
          "200": {
            "description": "Lista de películas obtenida exitosamente.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "integer"
                      },
                      "title": {
                        "type": "string"
                      },
                      "synopsis": {
                        "type": "string"
                      },
                      "duration": {
                        "type": "integer"
                      },
                      "director": {
                        "type": "string"
                      },
                      "isActive": {
                        "type": "boolean"
                      }
                    }
                  }
                },
                "example": [
                  {
                    "id": 1,
                    "title": "Inception",
                    "synopsis": "Un ladrón que roba secretos corporativos a través del uso de la tecnología de compartir sueños.",
                    "duration": 148,
                    "director": "Christopher Nolan",
                    "isActive": true
                  },
                  {
                    "id": 2,
                    "title": "Interstellar",
                    "synopsis": "Un equipo de exploradores viaja a través de un agujero de gusano en el espacio.",
                    "duration": 169,
                    "director": "Christopher Nolan",
                    "isActive": true
                  }
                ]
              }
            }
          },
          "500": {
            "description": "Error interno del servidor."
          }
        },
        "description": "### Obtener catálogo de películas activas\n\nLista todas las películas disponibles en el catálogo con proyección activa.\n\n- **Autenticación requerida:** No (Público).\n- **Respuesta:** Devuelve `success` y `data` con el array de películas (título, sinopsis, duración, clasificación, director)."
      }
    },
    "/movies/upcoming": {
      "get": {
        "summary": "Obtener las películas en estado \"Próximo Estreno\"",
        "tags": [
          "Movies"
        ],
        "responses": {
          "200": {
            "description": "Lista de próximos estrenos obtenida exitosamente.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "integer"
                      },
                      "title": {
                        "type": "string"
                      },
                      "posterUrl": {
                        "type": "string"
                      },
                      "releaseDate": {
                        "type": "string",
                        "format": "date"
                      },
                      "genres": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      },
                      "classification": {
                        "type": "string"
                      },
                      "duration": {
                        "type": "integer"
                      },
                      "trailerUrl": {
                        "type": "string"
                      },
                      "synopsis": {
                        "type": "string"
                      },
                      "daysUntil": {
                        "type": "integer"
                      }
                    }
                  }
                },
                "example": [
                  {
                    "id": 10,
                    "title": "Dune: Parte Tres",
                    "posterUrl": "https://images.unsplash.com/photo-1534447677768-be436bb09401",
                    "releaseDate": "2026-11-20",
                    "genres": [
                      "Ciencia Ficción",
                      "Aventura"
                    ],
                    "classification": "+12",
                    "duration": 165,
                    "trailerUrl": "https://www.youtube.com/watch?v=mock-dune3",
                    "synopsis": "Paul Atreides continúa su viaje épico liderando a los Fremen.",
                    "daysUntil": 60
                  }
                ]
              }
            }
          },
          "500": {
            "description": "Error interno del servidor."
          }
        },
        "description": "### Obtener películas en estado \"Próximo Estreno\"\n\nLista los títulos programados para estreno futuro en la plataforma.\n\n- **Autenticación requerida:** No (Público).\n- **Respuesta:** Devuelve `success` y `data` con las películas en próximo estreno."
      }
    },
    "/movies/upcoming/{id}": {
      "get": {
        "summary": "Obtener el detalle de una película en estado \"Próximo Estreno\"",
        "tags": [
          "Movies"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer",
              "example": 1
            },
            "description": "ID de la película."
          }
        ],
        "responses": {
          "200": {
            "description": "Detalle del próximo estreno obtenido exitosamente.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer"
                    },
                    "title": {
                      "type": "string"
                    },
                    "posterUrl": {
                      "type": "string"
                    },
                    "releaseDate": {
                      "type": "string",
                      "format": "date"
                    },
                    "genres": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "classification": {
                      "type": "string"
                    },
                    "duration": {
                      "type": "integer"
                    },
                    "trailerUrl": {
                      "type": "string"
                    },
                    "synopsis": {
                      "type": "string"
                    },
                    "daysUntil": {
                      "type": "integer"
                    }
                  }
                },
                "example": {
                  "id": 10,
                  "title": "Dune: Parte Tres",
                  "posterUrl": "https://images.unsplash.com/photo-1534447677768-be436bb09401",
                  "releaseDate": "2026-11-20",
                  "genres": [
                    "Ciencia Ficción",
                    "Aventura"
                  ],
                  "classification": "+12",
                  "duration": 165,
                  "trailerUrl": "https://www.youtube.com/watch?v=mock-dune3",
                  "synopsis": "Paul Atreides continúa su viaje épico liderando a los Fremen.",
                  "daysUntil": 60
                }
              }
            }
          },
          "400": {
            "description": "Id de la película inválido."
          },
          "404": {
            "description": "Próximo estreno no encontrado."
          },
          "500": {
            "description": "Error interno del servidor."
          }
        },
        "description": "### Obtener detalle de una película en \"Próximo Estreno\"\n\nConsulta la ficha completa de un título en próximo estreno.\n\n- **Parámetros de ruta:** `:id` (ID de la película).\n- **Respuesta:** Devuelve `success` y `data` con el detalle técnico de la película."
      }
    },
    "/movies/weekly": {
      "get": {
        "summary": "Obtener cartelera semanal (próximos 7 días) para la ubicación del usuario",
        "tags": [
          "Movies"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "cityId",
            "required": false,
            "schema": {
              "type": "integer",
              "example": 1
            },
            "description": "ID de la ciudad para filtrar la cartelera."
          }
        ],
        "responses": {
          "200": {
            "description": "Cartelera semanal obtenida exitosamente."
          },
          "500": {
            "description": "Error interno del servidor."
          }
        },
        "description": "### Obtener cartelera semanal (próximos 7 días)\n\nRetorna la cartelera de películas programadas para los próximos 7 días a partir de la ciudad del usuario (HU-003).\n\n- **Query Params:** `cityId` (Obligatorio, ID de la ciudad para geolocalización).\n- **Respuesta:** Devuelve `success` y `data` con las películas y sus respectivas fechas de proyección."
      }
    },
    "/movies/today": {
      "get": {
        "summary": "Obtener películas programadas para el día de hoy",
        "tags": [
          "Movies"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "cityId",
            "required": false,
            "schema": {
              "type": "integer",
              "example": 1
            },
            "description": "ID de la ciudad."
          }
        ],
        "responses": {
          "200": {
            "description": "Películas de hoy obtenidas exitosamente."
          },
          "500": {
            "description": "Error interno del servidor."
          }
        },
        "description": "### Obtener películas programadas para el día de hoy\n\nRetorna los títulos con funciones programadas para la fecha actual en la ciudad seleccionada.\n\n- **Query Params:** `cityId` (Obligatorio, ID de la ciudad).\n- **Respuesta:** Devuelve `success` y `data` con la cartelera del día."
      }
    },
    "/movies/filter": {
      "get": {
        "summary": "Filtrar películas/funciones por criterios (Fecha, Género, Idioma, Formato, etc.)",
        "tags": [
          "Movies"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "cityId",
            "required": false,
            "schema": {
              "type": "integer",
              "example": 1
            },
            "description": "ID de la ciudad."
          },
          {
            "in": "query",
            "name": "date",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date",
              "example": "2026-08-24"
            },
            "description": "Fecha a consultar (YYYY-MM-DD)."
          },
          {
            "in": "query",
            "name": "genre",
            "required": false,
            "schema": {
              "type": "string",
              "example": "Acción"
            },
            "description": "Género de la película."
          },
          {
            "in": "query",
            "name": "rating",
            "required": false,
            "schema": {
              "type": "string",
              "example": "+12"
            },
            "description": "Clasificación por edad."
          },
          {
            "in": "query",
            "name": "language",
            "required": false,
            "schema": {
              "type": "string",
              "example": "Doblada"
            },
            "description": "Idioma (Doblada/Subtitulada)."
          },
          {
            "in": "query",
            "name": "format",
            "required": false,
            "schema": {
              "type": "string",
              "example": "2D"
            },
            "description": "Formato de sala (2D, 3D, IMAX, VIP)."
          },
          {
            "in": "query",
            "name": "cinemaId",
            "required": false,
            "schema": {
              "type": "integer",
              "example": 1
            },
            "description": "Complejo/cine específico."
          },
          {
            "in": "query",
            "name": "available",
            "required": false,
            "schema": {
              "type": "string",
              "example": "true"
            },
            "description": "Filtro 'Disponible' para ocultar agotados (true/false)."
          }
        ],
        "responses": {
          "200": {
            "description": "Cartelera filtrada exitosamente."
          },
          "500": {
            "description": "Error interno del servidor."
          }
        },
        "description": "### Filtrar películas y funciones por criterios múltiples\n\nBúsqueda avanzada con filtros combinables: fecha, género, clasificación por edad, formato (2D, 3D, IMAX, VIP), idioma y complejo de cine.\n\n- **Query Params:** `cityId`, `date`, `genre`, `rating`, `language`, `format`, `cinemaId`, `available`.\n- **Respuesta:** Devuelve `success` y `data` con la lista filtrada de títulos y horarios disponibles."
      }
    },
    "/movies/{id}": {
      "get": {
        "summary": "Obtener el detalle completo de una película por su ID",
        "tags": [
          "Movies"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer",
              "example": 1
            },
            "description": "ID de la película."
          }
        ],
        "responses": {
          "200": {
            "description": "Detalle de película obtenido correctamente."
          },
          "404": {
            "description": "Película no encontrada."
          },
          "500": {
            "description": "Error interno del servidor."
          }
        },
        "description": "### Obtener detalle completo de una película por ID\n\nFicha técnica completa: título original, sinopsis, duración en minutos, género, director, clasificación y afiche promocional.\n\n- **Parámetros de ruta:** `:id` (ID de la película).\n- **Respuesta:** Devuelve `success` y `data` con el detalle de la película."
      }
    },
    "/movies/{id}/functions": {
      "get": {
        "summary": "Obtener las funciones activas y futuras para una película",
        "tags": [
          "Movies"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer",
              "example": 1
            },
            "description": "ID de la película."
          },
          {
            "in": "query",
            "name": "cityId",
            "required": false,
            "schema": {
              "type": "integer",
              "example": 1
            },
            "description": "ID de la ciudad para filtrar funciones."
          }
        ],
        "responses": {
          "200": {
            "description": "Funciones futuras obtenidas correctamente."
          },
          "500": {
            "description": "Error interno del servidor."
          }
        },
        "description": "### Obtener funciones activas y futuras de una película\n\nLista todas las funciones disponibles para la película indicada, con opción de filtro por ciudad.\n\n- **Parámetros de ruta:** `:id` (ID de la película).\n- **Query Params:** `cityId` (ID de la ciudad).\n- **Respuesta:** Devuelve `success` y `data` con las funciones asociadas."
      }
    },
    "/movies/{id}/recommendations": {
      "get": {
        "summary": "Obtener recomendaciones de películas similares",
        "tags": [
          "Movies"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer",
              "example": 1
            },
            "description": "ID de la película de referencia."
          }
        ],
        "responses": {
          "200": {
            "description": "Recomendaciones de películas obtenidas exitosamente."
          },
          "500": {
            "description": "Error interno del servidor."
          }
        },
        "description": "### Obtener recomendaciones de películas similares\n\nRetorna títulos recomendados basados en afinidad de género y director respecto a la película consultada.\n\n- **Parámetros de ruta:** `:id` (ID de la película de referencia).\n- **Respuesta:** Devuelve `success` y `data` con la lista de títulos afines."
      }
    },
...
    "/functions/movie/{id}": {
      "get": {
        "summary": "Obtener funciones de una película con filtros (formato, fecha, cine)",
        "tags": [
          "Functions"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer",
              "example": 1
            },
            "description": "ID de la película"
          },
          {
            "in": "query",
            "name": "format",
            "required": false,
            "schema": {
              "type": "string",
              "example": "2D"
            },
            "description": "Formato de proyección (2D, 3D, IMAX, VIP)"
          },
          {
            "in": "query",
            "name": "date",
            "required": false,
            "schema": {
              "type": "string",
              "example": "2026-08-24"
            },
            "description": "Fecha de la función (YYYY-MM-DD)"
          },
          {
            "in": "query",
            "name": "cinemaId",
            "required": false,
            "schema": {
              "type": "integer",
              "example": 1
            },
            "description": "ID del cine o complejo"
          }
        ],
        "responses": {
          "200": {
            "description": "Funciones obtenidas exitosamente"
          },
          "404": {
            "description": "Película no encontrada"
          }
        },
        "description": "### Obtener funciones de una película con filtros\n\nConsulta las funciones disponibles para una película según formato de proyección (2D, 3D, IMAX, VIP), fecha y complejo.\n\n- **Parámetros de ruta:** `:id` (ID de la película).\n- **Query Params:** `format`, `date`, `cinemaId`.\n- **Respuesta:** Devuelve `success` y `data` con el listado de funciones y salas."
      }
    },
...
  },
...
}